cpu:aarch64:bnorm_s8: fixed data type

+++ src/cpu/aarch64/jit_uni_batch_normalization_s8.cpp
@@ -39,7 +39,7 @@ namespace {
 
 using namespace Xbyak_aarch64;
 
-using data_t = uint32_t;
+using data_t = int8_t;
 
 struct call_params_t {
     // keep int sizes at 8 bytes -- jit code expects this
